Double_MAST
This script is designed to plot the overall coverage of markers along each chromosome, but also highlight specific markers of interest, allowing the distribution and relative marker of interest locations to be more clearly examined. It can be quite useful for assessing the quality of a given marker platform, and has a number of other uses.